+++ routes.js
@@ -1,12 +1,11 @@
-var random   = require( "keigai" ).util.number.random,
-    response = require( "./lib/tenso" ).response;
+var uuid  = require( "keigai" ).util.uuid;
 
 module.exports.get = {
 	"/": [],
 	"/reports/tps": function ( req, res ) {
-		this.respond( req, res, response( new Error( "TPS Cover Sheet not attached" ), 785 ), 785 );
+		this.respond( req, res, new Error( "TPS Cover Sheet not attached" ), 785 );
 	},
-	"/random": function ( req, res ) {
-		this.respond( req, res, response( random() ) );
+	"/uuid": function ( req, res ) {
+		this.respond( req, res, uuid() );
 	}
 }