+++ src/constructor.js
@@ -8,3 +8,28 @@ function Tenso () {
 	this.server   = new TurtleIO();
 	this.version  = "{{VERSION}}";
 }
+
+/**
+ * Setting constructor loop
+ *
+ * @method constructor
+ * @memberOf Tenso
+ * @type {Function}
+ */
+Tenso.prototype.constructor = Tenso;
+
+/**
+ * Sends a response to the Client
+ *
+ * @method respond
+ * @memberOf Tenso
+ * @param  {Object} req     Client request
+ * @param  {Object} res     Client response
+ * @param  {Mixed}  arg     Response body
+ * @param  {Number} status  Response status
+ * @param  {Object} headers Response headers
+ * @return {Undefined}      undefined
+ */
+Tenso.prototype.respond = function ( req, res, arg, status, headers ) {
+	this.server.respond( req, res, response( arg, status ), status, headers );
+};