+++ src/intro.js
@@ -6,6 +6,7 @@ var TurtleIO = require( "turtle.io" ),
     CONFIG   = require( __dirname + "/../config.json" ),
     keigai   = require( "keigai" ),
     util     = keigai.util,
+    array    = util.array,
     clone    = util.clone,
     iterate  = util.iterate,
     merge    = util.merge;