+++ src/outro.js
@@ -1,4 +1,2 @@
-module.exports.factory  = factory;
-module.exports.prepare  = prepare;
-module.exports.response = response;
+module.exports = factory;
 } )();