+++ src/factory.js
@@ -9,7 +9,7 @@ function factory ( arg ) {
 	var HOSTNAME = arg ? arg.hostname || "localhost" : "localhost",
         vhosts   = {},
         config   = arg ? merge( clone( CONFIG, true ), arg ) : CONFIG,
-        instance;
+        auth, instance;
 
 	if ( !config.port ) {
 		console.error( "Invalid configuration" );
@@ -20,6 +20,16 @@ function factory ( arg ) {
 	config.root       = __dirname + "/../";
 	config.vhosts     = vhosts;
 	config["default"] = HOSTNAME;
+
+	if ( config.auth !== null ) {
+		auth = {};
+		auth[HOSTNAME] = {
+			authRealm : config.auth.realm || "Private",
+			authList  : config.auth.list  || config.auth
+		};
+
+		config.auth = auth;
+	}
 
 	instance = new Tenso();