+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 
 ## 0.0.3
-- Created `response()` & added it to the exports
+- Updated `turtle.io` to 2.1.9 to gain customization of JSON formatting
+- Updated routing such that handlers execute with the context of the Tensō instance
+- Created `Tenso.prototype.respond()` to provide a very simple method for creating a response from custom routes
+- Updated README with more content
+- Updated default HTML page when no routes are loaded
 
 ## 0.0.2
 - Initial implementation of routing via external module